user show page

+++ db/migrate/20200706144834_create_donations.rb
@@ -3,7 +3,7 @@ class CreateDonations < ActiveRecord::Migration[6.0]
     create_table :donations do |t|
       t.integer :user_id
       t.integer :pet_id
-      t.float :donation_amount
+      t.integer :donation_amount
       t.datetime :date
 
       t.timestamps
